tgbot/requirements.txt
aiogram==2.20
aioredis==2.0
Jinja2==3.1.2
loguru==0.6.0
pyyaml==6.0
requests==2.28.0